+++ lib/settings/pages/gesture_settings_page.dart
@@ -25,8 +25,8 @@ class _GestureSettingsPageState extends State<GestureSettingsPage> {
   bool enablePostGestures = true;
   SwipeAction leftPrimaryPostGesture = SwipeAction.upvote;
   SwipeAction leftSecondaryPostGesture = SwipeAction.downvote;
-  SwipeAction rightPrimaryPostGesture = SwipeAction.reply;
-  SwipeAction rightSecondaryPostGesture = SwipeAction.save;
+  SwipeAction rightPrimaryPostGesture = SwipeAction.save;
+  SwipeAction rightSecondaryPostGesture = SwipeAction.toggleRead;
 
   /// Comment Gestures
   bool enableCommentGestures = true;
@@ -43,7 +43,6 @@ class _GestureSettingsPageState extends State<GestureSettingsPage> {
     ListPickerItem(icon: Icons.north_rounded, label: SwipeAction.upvote.label, payload: SwipeAction.upvote),
     ListPickerItem(icon: Icons.south_rounded, label: SwipeAction.downvote.label, payload: SwipeAction.downvote),
     ListPickerItem(icon: Icons.star_outline_rounded, label: SwipeAction.save.label, payload: SwipeAction.save),
-    ListPickerItem(icon: Icons.reply_rounded, label: SwipeAction.reply.label, payload: SwipeAction.reply),
     ListPickerItem(icon: Icons.markunread_outlined, label: SwipeAction.toggleRead.label, payload: SwipeAction.toggleRead),
     ListPickerItem(icon: Icons.not_interested_rounded, label: SwipeAction.none.label, payload: SwipeAction.none),
   ];
@@ -133,8 +132,8 @@ class _GestureSettingsPageState extends State<GestureSettingsPage> {
       enablePostGestures = prefs.getBool('setting_gesture_enable_post_gestures') ?? true;
       leftPrimaryPostGesture = SwipeAction.values.byName(prefs.getString('setting_gesture_post_left_primary_gesture') ?? SwipeAction.upvote.name);
       leftSecondaryPostGesture = SwipeAction.values.byName(prefs.getString('setting_gesture_post_left_secondary_gesture') ?? SwipeAction.downvote.name);
-      rightPrimaryPostGesture = SwipeAction.values.byName(prefs.getString('setting_gesture_post_right_primary_gesture') ?? SwipeAction.reply.name);
-      rightSecondaryPostGesture = SwipeAction.values.byName(prefs.getString('setting_gesture_post_right_secondary_gesture') ?? SwipeAction.save.name);
+      rightPrimaryPostGesture = SwipeAction.values.byName(prefs.getString('setting_gesture_post_right_primary_gesture') ?? SwipeAction.save.name);
+      rightSecondaryPostGesture = SwipeAction.values.byName(prefs.getString('setting_gesture_post_right_secondary_gesture') ?? SwipeAction.toggleRead.name);
 
       // Comment Gestures
       enableCommentGestures = prefs.getBool('setting_gesture_enable_comment_gestures') ?? true;