+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Changed to default feed type to be "All" rather than "Local" - contribution from @CTalvio
 - Optimization improvements to comment cards and calculating published/edited time - contribution from @ajsosa
 - Improved UI navigation experience when logging in - contribution from @micahmo
+- Posts no longer have the reply swipe gesture
 
 ### Fixed
 - Fixed issue where comment thread would show spinning indicator even after all comments have been loaded - contribution from @ajsosa